Refactor API routes to use repository functions for data fetching
