+++ word-chain/app/api/leaderboard/route.ts
@@ -1,18 +1,7 @@
 import { NextResponse } from "next/server";
-
-type Leader = { user: string; score: number; streak: number; wins: number };
-
-const LEADERBOARD: Leader[] = [
-  { user: "alice", score: 1240, streak: 12, wins: 5 },
-  { user: "bob", score: 1180, streak: 8, wins: 3 },
-  { user: "chris", score: 990, streak: 4, wins: 2 }
-];
-
-const TODAY = { date: "2025-08-26", winners: ["alice"], fastest: [{ user: "alice", timeSec: 72 }] };
+import { getLeaders, getTodaySummary } from "@/lib/repo/leaderboard";
 
 export async function GET() {
-  return NextResponse.json(
-    { leaders: LEADERBOARD, today: TODAY },
-    { headers: { "Cache-Control": "no-store" } }
-  );
+  const [leaders, today] = await Promise.all([getLeaders(), getTodaySummary()]);
+  return NextResponse.json({ leaders, today }, { headers: { "Cache-Control": "no-store" } });
 }
