+++ word-chain/app/api/validate/route.ts
@@ -1,22 +1,9 @@
 import { NextRequest, NextResponse } from "next/server";
-
-const DICTIONARY = new Set<string>([
-  "cold",
-  "cord",
-  "card",
-  "ward",
-  "warm",
-  "head",
-  "heal",
-  "teal",
-  "tell",
-  "tall",
-  "tail"
-]);
+import { validateWord } from "@/lib/repo/validate";
 
 export async function POST(req: NextRequest) {
   const body = (await req.json()) as { word: string };
   const word = (body?.word || "").toLowerCase();
-  const valid = DICTIONARY.has(word);
+  const valid = await validateWord(word);
   return NextResponse.json({ valid });
 }
